Removing date-of-birth and updating applicant-information.
