+++ frontend/app/routes/public/apply/$id/adult/applicant-information.tsx
@@ -1,5 +1,3 @@
-import { useMemo } from 'react';
-
 import { data, redirect, useFetcher } from 'react-router';
 
 import { faChevronLeft, faChevronRight } from '@fortawesome/free-solid-svg-icons';
@@ -11,21 +9,21 @@ import type { Route } from './+types/applicant-information';
 
 import { TYPES } from '~/.server/constants';
 import { loadApplyAdultState } from '~/.server/routes/helpers/apply-adult-route-helpers';
-import type { ApplicantInformationState } from '~/.server/routes/helpers/apply-route-helpers';
-import { applicantInformationStateHasPartner, getAgeCategoryFromDateString, saveApplyState } from '~/.server/routes/helpers/apply-route-helpers';
-import { getFixedT, getLocale } from '~/.server/utils/locale.utils';
+import { getAgeCategoryFromDateString, saveApplyState } from '~/.server/routes/helpers/apply-route-helpers';
+import { getFixedT } from '~/.server/utils/locale.utils';
 import { transformFlattenedError } from '~/.server/utils/zod.utils';
 import { Button, ButtonLink } from '~/components/buttons';
 import { Collapsible } from '~/components/collapsible';
 import { CsrfTokenInput } from '~/components/csrf-token-input';
+import { DatePickerField } from '~/components/date-picker-field';
 import { useErrorSummary } from '~/components/error-summary';
 import { InputPatternField } from '~/components/input-pattern-field';
-import type { InputRadiosProps } from '~/components/input-radios';
-import { InputRadios } from '~/components/input-radios';
 import { InputSanitizeField } from '~/components/input-sanitize-field';
 import { LoadingButton } from '~/components/loading-button';
 import { Progress } from '~/components/progress';
+import { useCurrentLanguage } from '~/hooks';
 import { pageIds } from '~/page-ids';
+import { extractDateParts, getAgeFromDateString, isPastDateString, isValidDateString } from '~/utils/date-utils';
 import { getTypedI18nNamespaces } from '~/utils/locale-utils';
 import { mergeMeta } from '~/utils/meta-utils';
 import type { RouteHandleData } from '~/utils/route-utils';
@@ -53,15 +51,11 @@ export const meta: Route.MetaFunction = mergeMeta(({ data }) => {
 export async function loader({ context: { appContainer, session }, params, request }: Route.LoaderArgs) {
   const state = loadApplyAdultState({ params, request, session });
   const t = await getFixedT(request, handle.i18nNamespaces);
-  const locale = getLocale(request);
-  const maritalStatuses = appContainer.get(TYPES.domain.services.MaritalStatusService).listLocalizedMaritalStatuses(locale);
 
   const meta = { title: t('gcweb:meta.title.template', { title: t('apply-adult:applicant-information.page-title') }) };
 
-  invariant(state.dateOfBirth, 'Expected state.dateOfBirth to be defined');
-  const ageCategory = getAgeCategoryFromDateString(state.dateOfBirth);
-
-  return { ageCategory, defaultState: state.applicantInformation, editMode: state.editMode, id: state.id, maritalStatuses, meta };
+  // TODO: Once the state has been reworked, we can change 'defaultState' to 'defaultState: state.applicantInformation'
+  return { defaultState: state, editMode: state.editMode, id: state.id, meta };
 }
 
 export async function action({ context: { appContainer, session }, params, request }: Route.ActionArgs) {
@@ -77,71 +71,96 @@ export async function action({ context: { appContainer, session }, params, reque
 
   if (formAction === FORM_ACTION.cancel) {
     invariant(state.applicantInformation, 'Expected state.applicantInformation to be defined');
-
-    if (applicantInformationStateHasPartner(state.applicantInformation) && state.partnerInformation === undefined) {
-      const errorMessage = t('apply-adult:applicant-information.error-message.marital-status-no-partner-information');
-      const flattenedErrors: z.typeToFlattenedError<ApplicantInformationState> = { formErrors: [errorMessage], fieldErrors: { maritalStatus: [errorMessage] } };
-      return { errors: transformFlattenedError(flattenedErrors) };
-    }
-
     return redirect(getPathById('public/apply/$id/adult/review-information', params));
   }
 
   // Form action Continue & Save
   // state validation schema
-  const applicantInformationSchema = z.object({
-    socialInsuranceNumber: z
-      .string()
-      .trim()
-      .min(1, t('apply-adult:applicant-information.error-message.sin-required'))
-      .superRefine((sin, ctx) => {
-        if (!isValidSin(sin)) {
-          ctx.addIssue({ code: z.ZodIssueCode.custom, message: t('apply-adult:applicant-information.error-message.sin-valid') });
-        } else if (state.partnerInformation && formatSin(sin) === formatSin(state.partnerInformation.socialInsuranceNumber)) {
-          ctx.addIssue({ code: z.ZodIssueCode.custom, message: t('apply-adult:applicant-information.error-message.sin-unique') });
-        }
-      }),
-    firstName: z
-      .string()
-      .trim()
-      .min(1, t('apply-adult:applicant-information.error-message.first-name-required'))
-      .max(100)
-      .refine(isAllValidInputCharacters, t('apply-adult:applicant-information.error-message.characters-valid'))
-      .refine((firstName) => !hasDigits(firstName), t('apply-adult:applicant-information.error-message.first-name-no-digits')),
-    lastName: z
-      .string()
-      .trim()
-      .min(1, t('apply-adult:applicant-information.error-message.last-name-required'))
-      .max(100)
-      .refine(isAllValidInputCharacters, t('apply-adult:applicant-information.error-message.characters-valid'))
-      .refine((lastName) => !hasDigits(lastName), t('apply-adult:applicant-information.error-message.last-name-no-digits')),
-    maritalStatus: z
-      .string({ errorMap: () => ({ message: t('apply-adult:applicant-information.error-message.marital-status-required') }) })
-      .trim()
-      .min(1, t('apply-adult:applicant-information.error-message.marital-status-required')),
-  }) satisfies z.ZodType<ApplicantInformationState>;
+  const applicantInformationSchema = z
+    .object({
+      socialInsuranceNumber: z
+        .string()
+        .trim()
+        .min(1, t('apply-adult:applicant-information.error-message.sin-required'))
+        .superRefine((sin, ctx) => {
+          if (!isValidSin(sin)) {
+            ctx.addIssue({ code: z.ZodIssueCode.custom, message: t('apply-adult:applicant-information.error-message.sin-valid') });
+          } else if (state.partnerInformation && formatSin(sin) === formatSin(state.partnerInformation.socialInsuranceNumber)) {
+            ctx.addIssue({ code: z.ZodIssueCode.custom, message: t('apply-adult:applicant-information.error-message.sin-unique') });
+          }
+        }),
+      firstName: z
+        .string()
+        .trim()
+        .min(1, t('apply-adult:applicant-information.error-message.first-name-required'))
+        .max(100)
+        .refine(isAllValidInputCharacters, t('apply-adult:applicant-information.error-message.characters-valid'))
+        .refine((firstName) => !hasDigits(firstName), t('apply-adult:applicant-information.error-message.first-name-no-digits')),
+      lastName: z
+        .string()
+        .trim()
+        .min(1, t('apply-adult:applicant-information.error-message.last-name-required'))
+        .max(100)
+        .refine(isAllValidInputCharacters, t('apply-adult:applicant-information.error-message.characters-valid'))
+        .refine((lastName) => !hasDigits(lastName), t('apply-adult:applicant-information.error-message.last-name-no-digits')),
+      dateOfBirthYear: z.number({ required_error: t('applicant-information.error-message.date-of-birth-year-required'), invalid_type_error: t('applicant-information.error-message.date-of-birth-year-number') }),
+      dateOfBirthMonth: z.number({ required_error: t('applicant-information.error-message.date-of-birth-month-required') }),
+      dateOfBirthDay: z.number({ required_error: t('applicant-information.error-message.date-of-birth-day-required'), invalid_type_error: t('applicant-information.error-message.date-of-birth-day-number') }),
+      dateOfBirth: z.string(),
+    })
+    .superRefine((val, ctx) => {
+      // At this point the year, month and day should have been validated as positive integer
+      const dateOfBirthParts = extractDateParts(`${val.dateOfBirthYear}-${val.dateOfBirthMonth}-${val.dateOfBirthDay}`);
+      const dateOfBirth = `${dateOfBirthParts.year}-${dateOfBirthParts.month}-${dateOfBirthParts.day}`;
+
+      if (!isValidDateString(dateOfBirth)) {
+        ctx.addIssue({ code: z.ZodIssueCode.custom, message: t('applicant-information.error-message.date-of-birth-valid'), path: ['dateOfBirth'] });
+      } else if (!isPastDateString(dateOfBirth)) {
+        ctx.addIssue({ code: z.ZodIssueCode.custom, message: t('applicant-information.error-message.date-of-birth-is-past'), path: ['dateOfBirth'] });
+      } else if (getAgeFromDateString(dateOfBirth) > 150) {
+        ctx.addIssue({ code: z.ZodIssueCode.custom, message: t('applicant-information.error-message.date-of-birth-is-past-valid'), path: ['dateOfBirth'] });
+      }
+    })
+    .transform((val) => {
+      // At this point the year, month and day should have been validated as positive integer
+      const dateOfBirthParts = extractDateParts(`${val.dateOfBirthYear}-${val.dateOfBirthMonth}-${val.dateOfBirthDay}`);
+      return { ...val, dateOfBirth: `${dateOfBirthParts.year}-${dateOfBirthParts.month}-${dateOfBirthParts.day}` };
+    }); // satisfies z.ZodType<ApplicantInformationState>; TODO: Uncomment once the state has been reworked.
 
   const parsedDataResult = applicantInformationSchema.safeParse({
     socialInsuranceNumber: String(formData.get('socialInsuranceNumber') ?? ''),
     firstName: String(formData.get('firstName') ?? ''),
     lastName: String(formData.get('lastName') ?? ''),
-    maritalStatus: formData.get('maritalStatus') ? String(formData.get('maritalStatus')) : undefined,
+    dateOfBirthYear: formData.get('dateOfBirthYear') ? Number(formData.get('dateOfBirthYear')) : undefined,
+    dateOfBirthMonth: formData.get('dateOfBirthMonth') ? Number(formData.get('dateOfBirthMonth')) : undefined,
+    dateOfBirthDay: formData.get('dateOfBirthDay') ? Number(formData.get('dateOfBirthDay')) : undefined,
+    dateOfBirth: '',
   });
 
   if (!parsedDataResult.success) {
     return data({ errors: transformFlattenedError(parsedDataResult.error.flatten()) }, { status: 400 });
   }
 
-  const hasPartner = applicantInformationStateHasPartner(parsedDataResult.data);
-  const remove = !hasPartner ? 'partnerInformation' : undefined;
-  saveApplyState({ params, remove, session, state: { applicantInformation: parsedDataResult.data } });
+  saveApplyState({
+    params,
+    session,
+    state: {
+      applicantInformation: {
+        firstName: parsedDataResult.data.firstName,
+        lastName: parsedDataResult.data.lastName,
+        socialInsuranceNumber: parsedDataResult.data.socialInsuranceNumber,
+        maritalStatus: '', // TODO: remove once the state is reworked
+      },
+      dateOfBirth: parsedDataResult.data.dateOfBirth,
+    },
+  });
+
+  if (getAgeCategoryFromDateString(parsedDataResult.data.dateOfBirth) === 'youth') {
+    return redirect(getPathById('public/apply/$id/adult/living-independently', params));
+  }
 
   if (state.editMode) {
     return redirect(getPathById('public/apply/$id/adult/review-information', params));
-  }
-
-  if (hasPartner) {
-    return redirect(getPathById('public/apply/$id/adult/partner-information', params));
   }
 
   return redirect(getPathById('public/apply/$id/adult/contact-information', params));
@@ -149,7 +168,8 @@ export async function action({ context: { appContainer, session }, params, reque
 
 export default function ApplyFlowApplicationInformation({ loaderData, params }: Route.ComponentProps) {
   const { t } = useTranslation(handle.i18nNamespaces);
-  const { ageCategory, defaultState, editMode, maritalStatuses } = loaderData;
+  const { currentLanguage } = useCurrentLanguage();
+  const { defaultState, editMode } = loaderData;
 
   const fetcher = useFetcher<typeof action>();
   const isSubmitting = fetcher.state !== 'idle';
@@ -158,25 +178,12 @@ export default function ApplyFlowApplicationInformation({ loaderData, params }: 
   const errorSummary = useErrorSummary(errors, {
     firstName: 'first-name',
     lastName: 'last-name',
+    ...(currentLanguage === 'fr'
+      ? { dateOfBirth: 'date-picker-date-of-birth-day', dateOfBirthDay: 'date-picker-date-of-birth-day', dateOfBirthMonth: 'date-picker-date-of-birth-month' }
+      : { dateOfBirth: 'date-picker-date-of-birth-month', dateOfBirthMonth: 'date-picker-date-of-birth-month', dateOfBirthDay: 'date-picker-date-of-birth-day' }),
+    dateOfBirthYear: 'date-picker-date-of-birth-year',
     socialInsuranceNumber: 'social-insurance-number',
-    maritalStatus: 'input-radio-marital-status-option-0',
   });
-
-  function getBackButtonRouteId() {
-    if (ageCategory === 'adults') {
-      return 'public/apply/$id/adult/disability-tax-credit';
-    }
-
-    if (ageCategory === 'youth') {
-      return 'public/apply/$id/adult/living-independently';
-    }
-
-    return 'public/apply/$id/adult/date-of-birth';
-  }
-
-  const maritalStatusOptions = useMemo<InputRadiosProps['options']>(() => {
-    return maritalStatuses.map((status) => ({ defaultChecked: status.id === defaultState?.maritalStatus, children: status.name, value: status.id }));
-  }, [defaultState?.maritalStatus, maritalStatuses]);
 
   return (
     <>
@@ -204,7 +211,7 @@ export default function ApplyFlowApplicationInformation({ loaderData, params }: 
                 aria-description={t('applicant-information.name-instructions')}
                 autoComplete="given-name"
                 errorMessage={errors?.firstName}
-                defaultValue={defaultState?.firstName ?? ''}
+                defaultValue={defaultState.applicantInformation?.firstName ?? ''}
                 required
                 disableScreenReaderErrors
               />
@@ -216,12 +223,20 @@ export default function ApplyFlowApplicationInformation({ loaderData, params }: 
                 maxLength={100}
                 aria-description={t('applicant-information.name-instructions')}
                 autoComplete="family-name"
-                defaultValue={defaultState?.lastName ?? ''}
+                defaultValue={defaultState.applicantInformation?.lastName ?? ''}
                 errorMessage={errors?.lastName}
                 required
                 disableScreenReaderErrors
               />
             </div>
+            <DatePickerField
+              id="date-of-birth"
+              names={{ day: 'dateOfBirthDay', month: 'dateOfBirthMonth', year: 'dateOfBirthYear' }}
+              defaultValue={defaultState.dateOfBirth ?? ''}
+              legend={t('applicant-information.dob')}
+              errorMessages={{ all: errors?.dateOfBirth, year: errors?.dateOfBirthYear, month: errors?.dateOfBirthMonth, day: errors?.dateOfBirthDay }}
+              required
+            />
             <InputPatternField
               id="social-insurance-number"
               name="socialInsuranceNumber"
@@ -230,12 +245,12 @@ export default function ApplyFlowApplicationInformation({ loaderData, params }: 
               inputMode="numeric"
               helpMessagePrimary={t('apply-adult:applicant-information.help-message.sin')}
               helpMessagePrimaryClassName="text-black"
-              defaultValue={defaultState?.socialInsuranceNumber ?? ''}
+              defaultValue={defaultState.applicantInformation?.socialInsuranceNumber ?? ''}
               errorMessage={errors?.socialInsuranceNumber}
               required
               disableScreenReaderErrors
             />
-            <InputRadios id="marital-status" name="maritalStatus" legend={t('applicant-information.marital-status')} options={maritalStatusOptions} errorMessage={errors?.maritalStatus} required disableScreenReaderErrors />
+            {/* <InputRadios id="marital-status" name="maritalStatus" legend={t('applicant-information.marital-status')} options={maritalStatusOptions} errorMessage={errors?.maritalStatus} required disableScreenReaderErrors /> */}
           </div>
           {editMode ? (
             <div className="flex flex-wrap items-center gap-3">
@@ -259,7 +274,14 @@ export default function ApplyFlowApplicationInformation({ loaderData, params }: 
               >
                 {t('apply-adult:applicant-information.continue-btn')}
               </LoadingButton>
-              <ButtonLink id="back-button" routeId={getBackButtonRouteId()} params={params} disabled={isSubmitting} startIcon={faChevronLeft} data-gc-analytics-customclick="ESDC-EDSC:CDCP Online Application Form-Adult:Back - Applicant information click">
+              <ButtonLink
+                id="back-button"
+                routeId={`public/apply/$id/type-application`}
+                params={params}
+                disabled={isSubmitting}
+                startIcon={faChevronLeft}
+                data-gc-analytics-customclick="ESDC-EDSC:CDCP Online Application Form-Adult:Back - Applicant information click"
+              >
                 {t('apply-adult:applicant-information.back-btn')}
               </ButtonLink>
             </div>
