--- frontend/app/routes/public/apply/$id/adult/date-of-birth.tsx
@@ -1,189 +1,0 @@
-import { data, redirect, useFetcher } from 'react-router';
-
-import { faChevronLeft, faChevronRight } from '@fortawesome/free-solid-svg-icons';
-import { useTranslation } from 'react-i18next';
-import { z } from 'zod';
-
-import type { Route } from './+types/date-of-birth';
-
-import { TYPES } from '~/.server/constants';
-import { loadApplyAdultState } from '~/.server/routes/helpers/apply-adult-route-helpers';
-import { getAgeCategoryFromDateString, saveApplyState } from '~/.server/routes/helpers/apply-route-helpers';
-import { getFixedT } from '~/.server/utils/locale.utils';
-import { transformFlattenedError } from '~/.server/utils/zod.utils';
-import { Button, ButtonLink } from '~/components/buttons';
-import { CsrfTokenInput } from '~/components/csrf-token-input';
-import { DatePickerField } from '~/components/date-picker-field';
-import { useErrorSummary } from '~/components/error-summary';
-import { LoadingButton } from '~/components/loading-button';
-import { Progress } from '~/components/progress';
-import { useCurrentLanguage } from '~/hooks';
-import { pageIds } from '~/page-ids';
-import { extractDateParts, getAgeFromDateString, isPastDateString, isValidDateString } from '~/utils/date-utils';
-import { getTypedI18nNamespaces } from '~/utils/locale-utils';
-import { mergeMeta } from '~/utils/meta-utils';
-import type { RouteHandleData } from '~/utils/route-utils';
-import { getPathById } from '~/utils/route-utils';
-import { getTitleMetaTags } from '~/utils/seo-utils';
-
-export const handle = {
-  i18nNamespaces: getTypedI18nNamespaces('apply-adult', 'apply', 'gcweb'),
-  pageIdentifier: pageIds.public.apply.adult.dateOfBirth,
-  pageTitleI18nKey: 'apply-adult:eligibility.date-of-birth.page-title',
-} as const satisfies RouteHandleData;
-
-export const meta: Route.MetaFunction = mergeMeta(({ data }) => {
-  return getTitleMetaTags(data.meta.title);
-});
-
-export async function loader({ context: { appContainer, session }, params, request }: Route.LoaderArgs) {
-  const state = loadApplyAdultState({ params, request, session });
-  const t = await getFixedT(request, handle.i18nNamespaces);
-
-  const meta = { title: t('gcweb:meta.title.template', { title: t('apply-adult:eligibility.date-of-birth.page-title') }) };
-
-  const { dateOfBirth } = state;
-  return { id: state.id, meta, defaultState: { dateOfBirth }, editMode: state.editMode };
-}
-
-export async function action({ context: { appContainer, session }, params, request }: Route.ActionArgs) {
-  const formData = await request.formData();
-
-  const securityHandler = appContainer.get(TYPES.routes.security.SecurityHandler);
-  securityHandler.validateCsrfToken({ formData, session });
-
-  const state = loadApplyAdultState({ params, request, session });
-  const t = await getFixedT(request, handle.i18nNamespaces);
-
-  const dateOfBirthSchema = z
-    .object({
-      dateOfBirthYear: z.number({ required_error: t('apply-adult:eligibility.date-of-birth.error-message.date-of-birth-year-required'), invalid_type_error: t('apply-adult:eligibility.date-of-birth.error-message.date-of-birth-year-number') }),
-      dateOfBirthMonth: z.number({ required_error: t('apply-adult:eligibility.date-of-birth.error-message.date-of-birth-month-required') }),
-      dateOfBirthDay: z.number({ required_error: t('apply-adult:eligibility.date-of-birth.error-message.date-of-birth-day-required'), invalid_type_error: t('apply-adult:eligibility.date-of-birth.error-message.date-of-birth-day-number') }),
-      dateOfBirth: z.string(),
-    })
-    .superRefine((val, ctx) => {
-      // At this point the year, month and day should have been validated as positive integer
-      const dateOfBirthParts = extractDateParts(`${val.dateOfBirthYear}-${val.dateOfBirthMonth}-${val.dateOfBirthDay}`);
-      const dateOfBirth = `${dateOfBirthParts.year}-${dateOfBirthParts.month}-${dateOfBirthParts.day}`;
-
-      if (!isValidDateString(dateOfBirth)) {
-        ctx.addIssue({ code: z.ZodIssueCode.custom, message: t('apply-adult:eligibility.date-of-birth.error-message.date-of-birth-valid'), path: ['dateOfBirth'] });
-      } else if (!isPastDateString(dateOfBirth)) {
-        ctx.addIssue({ code: z.ZodIssueCode.custom, message: t('apply-adult:eligibility.date-of-birth.error-message.date-of-birth-is-past'), path: ['dateOfBirth'] });
-      } else if (getAgeFromDateString(dateOfBirth) > 150) {
-        ctx.addIssue({ code: z.ZodIssueCode.custom, message: t('apply-adult:eligibility.date-of-birth.error-message.date-of-birth-is-past-valid'), path: ['dateOfBirth'] });
-      }
-    })
-    .transform((val) => {
-      // At this point the year, month and day should have been validated as positive integer
-      const dateOfBirthParts = extractDateParts(`${val.dateOfBirthYear}-${val.dateOfBirthMonth}-${val.dateOfBirthDay}`);
-      return { ...val, dateOfBirth: `${dateOfBirthParts.year}-${dateOfBirthParts.month}-${dateOfBirthParts.day}` };
-    });
-
-  const parsedDataResult = dateOfBirthSchema.safeParse({
-    dateOfBirthYear: formData.get('dateOfBirthYear') ? Number(formData.get('dateOfBirthYear')) : undefined,
-    dateOfBirthMonth: formData.get('dateOfBirthMonth') ? Number(formData.get('dateOfBirthMonth')) : undefined,
-    dateOfBirthDay: formData.get('dateOfBirthDay') ? Number(formData.get('dateOfBirthDay')) : undefined,
-    dateOfBirth: '',
-  });
-
-  if (!parsedDataResult.success) {
-    return data({ errors: transformFlattenedError(parsedDataResult.error.flatten()) }, { status: 400 });
-  }
-
-  const ageCategory = getAgeCategoryFromDateString(parsedDataResult.data.dateOfBirth);
-
-  saveApplyState({
-    params,
-    session,
-    state: { dateOfBirth: parsedDataResult.data.dateOfBirth, disabilityTaxCredit: ageCategory === 'adults' ? state.disabilityTaxCredit : undefined, livingIndependently: ageCategory === 'youth' ? state.livingIndependently : undefined },
-  });
-
-  if (state.editMode) {
-    return redirect(getPathById('public/apply/$id/adult/review-information', params));
-  }
-
-  if (ageCategory === 'adults') {
-    return redirect(getPathById('public/apply/$id/adult/disability-tax-credit', params));
-  }
-
-  if (ageCategory === 'youth') {
-    return redirect(getPathById('public/apply/$id/adult/living-independently', params));
-  }
-
-  if (ageCategory === 'children') {
-    return redirect(getPathById('public/apply/$id/adult/parent-or-guardian', params));
-  }
-
-  return redirect(getPathById('public/apply/$id/adult/applicant-information', params));
-}
-
-export default function ApplyFlowDateOfBirth({ loaderData, params }: Route.ComponentProps) {
-  const { currentLanguage } = useCurrentLanguage();
-  const { t } = useTranslation(handle.i18nNamespaces);
-  const { defaultState, editMode } = loaderData;
-
-  const fetcher = useFetcher<typeof action>();
-  const isSubmitting = fetcher.state !== 'idle';
-
-  const errors = fetcher.data?.errors;
-  const errorSummary = useErrorSummary(errors, {
-    ...(currentLanguage === 'fr'
-      ? { dateOfBirth: 'date-picker-date-of-birth-day', dateOfBirthDay: 'date-picker-date-of-birth-day', dateOfBirthMonth: 'date-picker-date-of-birth-month' }
-      : { dateOfBirth: 'date-picker-date-of-birth-month', dateOfBirthMonth: 'date-picker-date-of-birth-month', dateOfBirthDay: 'date-picker-date-of-birth-day' }),
-    dateOfBirthYear: 'date-picker-date-of-birth-year',
-  });
-
-  return (
-    <>
-      <div className="my-6 sm:my-8">
-        <Progress value={33} size="lg" label={t('apply:progress.label')} />
-      </div>
-      <div className="max-w-prose">
-        <p className="mb-6">{t('apply-adult:eligibility.date-of-birth.description')}</p>
-        <p className="mb-4 italic">{t('apply:required-label')}</p>
-        <errorSummary.ErrorSummary />
-        <fetcher.Form method="post" noValidate>
-          <CsrfTokenInput />
-          <div className="mb-6 space-y-4">
-            <DatePickerField
-              id="date-of-birth"
-              names={{ day: 'dateOfBirthDay', month: 'dateOfBirthMonth', year: 'dateOfBirthYear' }}
-              defaultValue={defaultState.dateOfBirth ?? ''}
-              legend={t('apply-adult:eligibility.date-of-birth.form-instructions')}
-              errorMessages={{ all: errors?.dateOfBirth, year: errors?.dateOfBirthYear, month: errors?.dateOfBirthMonth, day: errors?.dateOfBirthDay }}
-              required
-            />
-          </div>
-          {editMode ? (
-            <div className="mt-8 flex flex-wrap items-center gap-3">
-              <Button variant="primary" id="continue-button" disabled={isSubmitting} data-gc-analytics-customclick="ESDC-EDSC:CDCP Online Application Form-Adult:Save - Date of birth click">
-                {t('apply-adult:eligibility.date-of-birth.save-btn')}
-              </Button>
-              <ButtonLink id="back-button" routeId="public/apply/$id/adult/review-information" params={params} disabled={isSubmitting} data-gc-analytics-customclick="ESDC-EDSC:CDCP Online Application Form-Adult:Cancel - Date of birth click">
-                {t('apply-adult:eligibility.date-of-birth.cancel-btn')}
-              </ButtonLink>
-            </div>
-          ) : (
-            <div className="mt-8 flex flex-row-reverse flex-wrap items-center justify-end gap-3">
-              <LoadingButton variant="primary" id="continue-button" loading={isSubmitting} endIcon={faChevronRight} data-gc-analytics-customclick="ESDC-EDSC:CDCP Online Application Form-Adult:Continue - Date of birth click">
-                {t('apply-adult:eligibility.date-of-birth.continue-btn')}
-              </LoadingButton>
-              <ButtonLink
-                id="back-button"
-                routeId="public/apply/$id/type-application" // returning to type/application... This file will be removed when all routes are refactored.
-                params={params}
-                disabled={isSubmitting}
-                startIcon={faChevronLeft}
-                data-gc-analytics-customclick="ESDC-EDSC:CDCP Online Application Form-Adult:Back - Date of birth click"
-              >
-                {t('apply-adult:eligibility.date-of-birth.back-btn')}
-              </ButtonLink>
-            </div>
-          )}
-        </fetcher.Form>
-      </div>
-    </>
-  );
-}
